Construct Quad Tree

+++ src/com/leetcode/www/ConstructQuadTree.java
@@ -26,7 +26,7 @@ class Solution {
      * Runtime: 8 ms, faster than 6.00% of Java online submissions for Construct Quad Tree.
      * Memory Usage: 44.3 MB, less than 45.98% of Java online submissions for Construct Quad Tree.
      *
-     * the following link is similar like my proposed one (following code)
+     * the following link is similar like my proposed one (bottom up recursive as following code)
      * https://leetcode.com/problems/construct-quad-tree/discuss/286829/Java-recursive
      *
      * @param grid
@@ -40,19 +40,19 @@ class Solution {
             if (n == 1) {
                 return new Node(grid[0][0] == 1, true, null, null, null, null);
             } else {
-                int mid = (int)(n / 2);
+                int mid = n / 2;
                 Node topLeft = construct(getSubGrid(0, 0, grid));
                 Node topRight = construct(getSubGrid(0, mid, grid));
                 Node bottomLeft = construct(getSubGrid(mid, 0, grid));
                 Node bottomRight = construct(getSubGrid(mid, mid, grid));
                 if (topLeft.isLeaf && topRight.isLeaf && bottomLeft.isLeaf && bottomRight.isLeaf) {
                     if (topLeft.val == topRight.val && topRight.val == bottomLeft.val &&
-                            bottomLeft.val == bottomRight.val) {
+                            bottomLeft.val == bottomRight.val) { // sub grids are same
                         return new Node(topLeft.val, true, null, null, null, null);
-                    } else {
+                    } else { // sub grids are different
                         return new Node(false, false, topLeft, topRight, bottomLeft, bottomRight);
                     }
-                } else {
+                } else { // sub grids are different
                     return new Node(false, false, topLeft, topRight, bottomLeft, bottomRight);
                 }
             }
@@ -60,7 +60,7 @@ class Solution {
     }
 
     private int[][] getSubGrid(int row, int col, int[][] grid) {
-        int len = (int)(grid.length / 2);
+        int len = grid.length / 2;
         int[][] subGrid = new int[len][len];
         for (int i = 0; i < len; i++) {
             for (int j = 0; j < len; j++) {
